Fixed both reported bugs involving the file explorer
Added checks to make sure the file explorer is returning value outputs. One check ensures it is an image file, the other makes sure a null is not returned which happens when the explorer is closed without selecting a file.

Both checks might be able to be condensed down to one check and we might be able to add a GUI popup instead of just a System.out on check failure.